requirements.txt
ccxt>=4.2.70
pandas>=2.2.0
numpy>=1.26.0
python-dotenv>=1.0.1
pytz>=2024.1
ta>=0.11.0